% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\name{bab}
\alias{bab}
\title{Use BAB framework to create a prompt}
\usage{
bab(
  before,
  after,
  bridge,
  template = "Before: {before}. \\n\\n After: {after}. \\n\\n Bridge: {bridge}",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Use BAB framework to create a prompt
}
\examples{
bab(
 before = "We're struggling to get traffic to our website",
 after = "We want to be top 10 in SEO rankings in our niche in 90 days",
 bridge = "Develop a detailed plan for mentioning all the measures we should take also include list of top 20 keywords"
)
}
