% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\name{chain_of_destiny}
\alias{chain_of_destiny}
\title{Use Chain of Destiny framework to create a prompt}
\usage{
chain_of_destiny(
  instructions,
  recursion,
  benchmark,
  additional_guidelines,
  template =
    "Instructions: {instructions} \\n\\n Recursion: {recursion} \\n\\n Benchmark: {benchmark} \\n\\n Additional Guidelines: {additional_guidelines}",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Use Chain of Destiny framework to create a prompt
}
\examples{
chain_of_destiny(
  instructions = 'Here is a prompt I use to generate cold emails for my landing page creation agency: "I need a cold email idea that will engage my [ideal customer] with a unique perspective on [subject] and persuade them to take [desired action]"
You will generate increasingly improved versions of this prompt.',
recursion = "Repeat the following 2 steps 5 times.
Step 1. Identify 1-3 points from the initial output which are missing.
Step 2. Write a new, improved output of identical length which includes the missing points.",
benchmark = "Here is more information on what makes a good prompt:
- Clarity and Specificity
- Be Clear: The prompt should be easy to understand.
- Be Specific: Vague questions will often yield vague answers. The more specific you are, the better the AI can generate a targeted response.
- Open-Ended vs Closed-Ended
- Open-Ended: Use open-ended questions for comprehensive answers or when you want the AI to generate multiple ideas.
- Closed-Ended: Use yes/no or either/or questions when you need a straightforward answer.",
additional_guidelines = "Remember to repeat the process 5 times"
)
}
