% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\name{chain_of_thought}
\alias{chain_of_thought}
\title{Use Chain of Thought framework to create a prompt}
\usage{
chain_of_thought(
  instructions,
  template = "{instructions}. \\n\\n Let's think through it step-by-step.",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Use Chain of Thought framework to create a prompt
}
\examples{
chain_of_thought("How do I improve my sales calls? I've only got a 15\% close rate right now, and I think it's because I'm not selling the dream enough.")
}
