% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_copy.R
\name{copy_with_message}
\alias{copy_with_message}
\title{Copy to Clipboard with Cli Message}
\usage{
copy_with_message(x, message = "Copied to clipboard")
}
\arguments{
\item{x}{string to copy to clipboard}

\item{message}{message to display}
}
\description{
Copy to Clipboard with Cli Message
}
