% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fewshot.R
\name{fewshot}
\alias{fewshot}
\title{Fewshot formatted prompt}
\usage{
fewshot(
  text,
  instructions = "",
  examples = data.frame(),
  template = "Text: {text}\\nClassification: {label}",
  prompt_template = "{instructions}{examples}{input}",
  separator = "\\n\\n",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{prompt_template}{string to format the prompt}

\item{separator}{string to separate examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Fewshot formatted prompt
}
