% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_copy.R
\name{glue_copy}
\alias{glue_copy}
\title{Use Glue and Copy String}
\usage{
glue_copy(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .trim = TRUE,
  .copy_to_clipboard = TRUE,
  .copy_message = "Copied to clipboard"
)
}
\arguments{
\item{.copy_message}{}
}
\description{
Use Glue and Copy String
}
