% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\name{rodes}
\alias{rodes}
\title{Use RODES framework to create a prompt}
\usage{
rodes(
  role,
  objective,
  details,
  examples,
  sense_check,
  template =
    "Role: {role} \\n\\n Objective: {objective} \\n\\n Details: {details} \\n\\n Examples: {examples} \\n\\n Sense Check: {sense_check}",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Use RODES framework to create a prompt
}
\examples{
rodes(
role = "You are a seasoned marketing professional with 10 years of experience in the industry.",
objective = "Please give me a list of the top 10 marketing strategies that I should implement to increase my sales by 20\% in the next quarter.",
details = "- The tweet should be no longer than 280 characters.
  - Use powerful and persuasive language.
  - Do not include hashtags or emojis.",
examples = 'Here are some examples you should use to model your answer (note - these are not on my desired topic. I just want you to understand the frameworks and styles that work")
1."Credit card debt is killing your financial progress.
Heres how you get rid of it ASAP:"',
sense_check = "Do you understand the objective and the specific guidelines for this task?"
)
}
