% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameworks.R
\name{rtf}
\alias{rtf}
\title{Use RTF framework to create a prompt}
\usage{
rtf(
  role,
  task,
  format,
  template = "Act like a {role}. Give me a {task} in {format} format.",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{template}{string to format the examples}

\item{copy_to_clipboard}{boolean to copy to clipboard}
}
\description{
Use RTF framework to create a prompt
}
\examples{
rtf("life coach with 30 years of experience in mentoring", "plan to improve my work-life balance", "table")
}
