% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_prompt}
\alias{save_prompt}
\title{Save prompt to text file}
\usage{
save_prompt(prompt, path)
}
\arguments{
\item{prompt}{string with prompt to save to file}

\item{path}{full file path to save the prompt to}
}
\description{
Save prompt to text file
}
\examples{
\dontrun{
 save_prompt("This is a prompt", "file.txt")
}
}
